Create pull_request_template.md
Here's my initial draft for a PR template. Hopefully, this will (a) allow the reviewer to immediately have all the context she needs to focus on the important details and (b) get the folks that are pushing PRs thinking about justification for their choices as well as critical issues like test coverage.